fix tests that rely on ACLs on as not used on linux build agent
